+++ CHANGELOG.md
@@ -7,13 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+[//]: # (--8<-- [start:released])
+
+# [0.2.0] - 2026-04-24
+
 ### Added
 
 - `DEFAULT_SECTIONS` containing the default Keep a Changelog sections.
 - `sections` option to recognize custom sections after release headings.
 - `extend_sections` option for additional sections alongside the default ones.
-
-[//]: # (--8<-- [start:released])
 
 ## [0.1.0] - 2026-04-11
 
@@ -21,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Slugifier` for changelog-aware heading slugs.
 
-[Unreleased]: https://github.com/jo-phil/changelog-slugs/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/jo-phil/changelog-slugs/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/jo-phil/changelog-slugs/releases/tag/0.2.0
 [0.1.0]: https://github.com/jo-phil/changelog-slugs/releases/tag/0.1.0